refactor(learnings): slim learnings.ts to formatting-only
Move parseLearningContent to runner.ts where it is consumed.
learnings.ts now exports only formatting functions (formatLearningsForPrompt,
formatLearningsForPr). Update module-level comment, test imports, and
ARCHITECTURE.md to reflect the focused responsibility.
